embedding-enrichment/requirements.txt
umap-learn>=0.5.5
psycopg2-binary>=2.9.9
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0 